requirements.txt
Django==1.6.5
django-extensions==1.3.10
djangorestframework==2.3.14
python-social-auth==0.1.26
django-guardian==1.2.4
django-waffle==0.10
South==1.0
twilio==3.6.6
